refactor: move card components to the single folder
