+++ src/components/_index.ts
@@ -1,4 +1,4 @@
-export { SearchBar } from '@/components/SearchBar/SearchBar';
-export { CardList } from '~compos/cardList/CardList';
+export { CardList } from '~compos/cards/cardList/CardList';
 export { Header } from '~compos/header/Header';
 export { Layout } from '~compos/layout/Layout';
+export { SearchBar } from '~compos/SearchBar/SearchBar';
